+++ JsonToolsNppPlugin/Properties/AssemblyInfo.cs
@@ -28,5 +28,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("6.1.1.15")]
-[assembly: AssemblyFileVersion("6.1.1.15")]
+[assembly: AssemblyVersion("6.1.1.16")]
+[assembly: AssemblyFileVersion("6.1.1.16")]